+++ lib/core.dart
@@ -5,6 +5,8 @@ export 'src/core/digit.dart';
 export 'src/core/equal.dart';
 export 'src/core/len.dart';
 export 'src/core/num.dart';
+export 'src/core/pair.dart';
 export 'src/core/req.dart';
 export 'src/core/rules.dart';
+export 'src/core/to_val_obj.dart';
 export 'src/core/trim.dart';