+++ lib/src/net/req_email.dart
@@ -10,10 +10,10 @@ class ReqEmail {
   /// [blank] the blank field message; defaults to 'required email'.
   /// [mal] the malformed email message.
   ReqEmail({String? blank, String? mal})
-      : _reqEmail =
-            Rules([Req(blank: blank ?? 'required email'), Email(mal: mal)]);
+      : _reqEmail = Req.val(ToValObj(Email(mal: mal)),
+            blank: blank ?? 'required email');
 
-  final Rules<String> _reqEmail;
+  final ValObj _reqEmail;
 
   /// Valid — returns null — if [input] is a non-blank well-formed email;
   /// otherwise, returns the blank error message if [input] is null, or the